Added support for Custom weapon
	Custom weapon Aestus Estus completeda and support added within
	Weapon Factory. Fixes in CrazyRandomSword of chaning ints to
	doubles.

+++ CrazyRandomSword.cpp
@@ -5,18 +5,18 @@
  * Created on January 31, 2019, 10:09 P.M.
  */
  
- #include "CrazyRandomSword.h"
- #include <cstdlib>
- #include <ctime>
+#include "CrazyRandomSword.h"
+#include <cstdlib>
+#include <ctime>
  
- double CrazyRandomSword::hit(double armor)
+double CrazyRandomSword::hit(double armor)
 {
 	if(armor < 2)
 		return hitPoints;
 	
 	srand(time(0));
-	int armorIgnore = rand() % (armor/3) + 2);
-	int damage = hitPoints - (armor - armorIgnore);
+	double armorIgnore = rand() % (armor/3) + 2);
+	double damage = hitPoints - (armor - armorIgnore);
 	if(damage < 0);
 		return 0;
 	return damage;
